pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dummy-accordo-agent"
version = "0.1.0"
description = "Agentic workflow for drafting Accord Project templates - GSoC 2026 MVP"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Adarsh", email = "adarshh347@example.com" }
]
keywords = [
    "accord-project",
    "concerto",
    "llm",
    "agents",
    "crewai",
    "smart-contracts",
    "legal-tech"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "crewai>=0.28.0",
    "crewai-tools>=0.2.0",
    "langchain-groq>=0.1.0",
    "groq>=0.4.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
accordo = "src.cli.main:main"

[project.urls]
Homepage = "https://github.com/adarshh347/dummy-accordo-agent"
Repository = "https://github.com/adarshh347/dummy-accordo-agent.git"
Issues = "https://github.com/adarshh347/dummy-accordo-agent/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
